+++ src/main/java/racingcar/controller/RacingCarGame.java
@@ -1,4 +1,11 @@
-package racingcar;
+package racingcar.controller;
+
+import racingcar.domain.CarListManager;
+import racingcar.domain.GameResult;
+import racingcar.domain.RoundRecord;
+import racingcar.domain.dto.CarStatusDto;
+import racingcar.domain.model.Car;
+import racingcar.view.RacingCarView;
 
 import java.util.ArrayList;
 import java.util.List;
