+++ src/main/java/racingcar/domain/dto/CarStatusDto.java
@@ -1,4 +1,6 @@
-package racingcar;
+package racingcar.domain.dto;
+
+import racingcar.domain.model.Car;
 
 public class CarStatusDto {
