+++ src/main/java/racingcar/domain/model/Car.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.domain.model;
 
 import java.util.Objects;
